+++ examples/tui/tui.rs
@@ -1,9 +1,9 @@
-use gb_rs::{
-    cart::{get_cart_header, CartridgeData},
+use gb_rs::{ 
     gb::GbRs,
     joypad::{JoypadDirection, JoypadInput},
     ppu::{BKG_WIDTH, SCREEN_HEIGHT, SCREEN_WIDTH},
     tile::Tile,
+    util::VecCart,
 };
 use std::{
     io,
@@ -372,75 +372,6 @@ impl Widget for OamWidget<'_> {
         }
 
         outer_block.render(area, buf);
-    }
-}
-
-struct VecCart {
-    rom: Vec<u8>,
-    ram: Vec<u8>,
-    save_path: Option<String>,
-}
-
-impl VecCart {
-    pub fn from_slice(data: &[u8], save_dir: Option<&str>) -> Self {
-        let header = get_cart_header(data);
-
-        let rom = Vec::from(data);
-
-        if let Some(dir) = save_dir {
-            let file = dir.to_owned() + &header.title;
-            let ram = std::fs::read(file.clone());
-
-            let ram: Vec<u8> = if ram.is_ok() {
-                ram.unwrap()
-            } else {
-                vec![0; header.ram_size as usize]
-            };
-
-            assert_eq!(ram.len(), header.ram_size as usize);
-
-            Self {
-                rom,
-                ram,
-                save_path: Some(file),
-            }
-        } else {
-            let ram = vec![0; header.ram_size as usize];
-            Self {
-                rom,
-                ram,
-                save_path: None,
-            }
-        }
-    }
-}
-
-impl Drop for VecCart {
-    fn drop(&mut self) {
-        if let Some(file) = &self.save_path {
-            std::fs::write(file, &self.ram).expect("Unable to save game...");
-        }
-    }
-}
-
-impl CartridgeData for VecCart {
-    type Rom = Vec<u8>;
-    type Ram = Vec<u8>;
-
-    fn rom(&self) -> &Self::Rom {
-        &self.rom
-    }
-
-    fn rom_mut(&mut self) -> &mut Self::Rom {
-        &mut self.rom
-    }
-
-    fn ram(&self) -> &Self::Ram {
-        &self.ram
-    }
-
-    fn ram_mut(&mut self) -> &mut Self::Ram {
-        &mut self.ram
     }
 }
