+++ src/lib.rs
@@ -1,4 +1,7 @@
-#![cfg_attr(not(test), no_std)]
+#![no_std]
+
+#[cfg(any(test, feature = "std"))]
+extern crate std;
 
 pub mod bus;
 pub mod cart;
@@ -10,3 +13,6 @@ pub mod oam;
 pub mod ppu;
 pub mod tile;
 pub mod timer;
+
+#[cfg(any(test, feature = "std"))]
+pub mod util;
